pyproject.toml
[project]
name = "regions-counter"

version = "1.0.0"

dependencies = [
    "pyshp",
    "pyproj",
    "mysql.connector",
    "requests",
    "shapely"
]

[tool.setuptools]
py-modules = []